requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
crispy-bootstrap5==2023.10
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.1
django-crispy-forms==2.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.6.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
